clients annotated with @BootstrapContextCompatible
