+++ aws-sdk-v2/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    api platform("software.amazon.awssdk:bom:2.16.59")
+    api platform("software.amazon.awssdk:bom:2.16.63")
     api project(":aws-common")
 
     compileOnly "org.graalvm.nativeimage:svm"
