--- aws-sdk-v2/src/main/java/io/micronaut/aws/sdk/v2/client/SdkHttpClientConfiguration.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2017-2021 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.aws.sdk.v2.client;
-
-/**
- * Configuration about {@link software.amazon.awssdk.http.SdkHttpClient}.
- * @author Sergio del Amo
- * @since 2.7.0
- */
-public interface SdkHttpClientConfiguration {
-
-    /**
-     *
-     * @return Whether the sdk http client should be loaded into the Bootstrap Context.
-     */
-    boolean getBootstrap();
-}
