+++ aws-sdk-v2/src/main/java/io/micronaut/aws/sdk/v2/client/apache/ApacheClientFactory.java
@@ -15,12 +15,11 @@
  */
 package io.micronaut.aws.sdk.v2.client.apache;
 
-import io.micronaut.aws.sdk.v2.client.SdkHttpClientConfigurationProperties;
 import io.micronaut.aws.sdk.v2.client.urlConnection.UrlConnectionClientFactory;
 import io.micronaut.context.annotation.Bean;
+import io.micronaut.context.annotation.BootstrapContextCompatible;
 import io.micronaut.context.annotation.Factory;
 import io.micronaut.context.annotation.Requires;
-import io.micronaut.core.util.StringUtils;
 import software.amazon.awssdk.http.SdkHttpClient;
 
 import javax.inject.Singleton;
@@ -31,8 +30,8 @@ import javax.inject.Singleton;
  * @author Álvaro Sánchez-Mariscal
  * @since 2.0.0
  */
+@BootstrapContextCompatible
 @Factory
-@Requires(property = SdkHttpClientConfigurationProperties.PREFIX + ".bootstrap", value = StringUtils.FALSE, defaultValue = StringUtils.FALSE)
 public class ApacheClientFactory {
 
     public static final String APACHE_SDK_HTTP_SERVICE = "software.amazon.awssdk.http.apache.ApacheSdkHttpService";
